Crash Fix. Working Fix. Couple of TO-DOs
It does work. There is a few things that i tested that dealt with consistent results.
Test Area is Rage Fire Chasm due to my famliarity of the dungeon as my personal test grounds in all wow revisions. And Moltencore Raid per @Jinnaix guidence
So far it is only tested with DUNGEON (Rage Fire Chasm) and RAID (Moltencore).

You MUST be half way thru the dungeon then kill a boss for it to trigger the save resurrection spot thru out the instance. if you die near the entrance upon entry you will spawn in a grave yard, but if you are half way thru the instances like i was with rage fire chasm and killed Taragaram the Hungerer or further then it will start saving the points for anything after you kill in that instance even by the entrance of the instance.
Same thing with Molten core, the dynamic resurrection didnt occur until you reached magmus and kill him.
TO DO:
5 Man or More  Return False Function. Where if we have a group of 5 or in a instance, the module will not work and will send the dead player outside of the entrance.
Raids, Figure out how to Ressurect players at the outside entrance of raids as it doesnt do it but acts like normal dungeon when doing a dynamic ressurection.

Co-Authored-By: Jinnaix <37972361+Jinnaix@users.noreply.github.com>