fixed switch around.
If player died in combat, released on corpse if raid and if they died outside of combat they released on entrance. That was wrong. switched them around. But we need to fix the Dynamic Resurrection.cpp in the scripts\custom folder. The Void for on entrance is what is wrong now for the auto tele to the last killed boss. need to figure out how to rewrite that if raid entrance is on.