Add: (Module): Solocraft Class Balance Modifier
With class scaling you got the new configs to adjust the individual class to use a certain precentage of the established solocraft configs. Single value does not cover scaling perfectly, at times it will make select classes over power. Class configurations are adjustable INT between 0 to 100. Representing 0 for zero scaling for the individual class to 100 for full scaling of the individual class. It isnt a float so dont start using decimals for the Class Balancing Modifier