feat (New Detection) Ignore Z-Axis
Thank you for the amazing input and information @mdX7 .
This is the ignore Z-Axis Detection.  The cheat is when a user blocks their z-axis from updating with the server resulting in them running\moving in the air. Flight Hack Detection would not detect them.
Warning: Some light false positives may occur on certain parts of the area\map if there is any "pot holes" of sorts. Like it is in Az-Core.

Co-Authored-By: ModoX <moardox@gmail.com>