Further Reduce False Positives
This is provided by @sveN295 and @Jinnaix who both are better people then me and for that I thank them for their willingness to assist in updating this ancient anticheat.
Per their notes:
We remove the added speed for jumping because otherwise permanently jumping doubles your allowed speed.

We check the last MovementInfo for the falling flag since falling down a hill and sliding a bit triggered a false positive

Co-Authored-By: Sven <36102838+sveN295@users.noreply.github.com>
Co-Authored-By: Jinnaix <37972361+Jinnaix@users.noreply.github.com>