Updated; Update SQL improved Column Counting
Removed enum DetectionTypes, it was based on bitmasking that i didnt think much of when i was adding on to it but we are not using bitmasking for option enabling, we are not embedded engineers, a slight moment jackpoz mentiond in the discord when i mentioned the bitmasking before being used. Column counting added in place for teleportations\ignore control\ignore z-axis since propper count is now done and handled. added exact force counting for speed hack as well. Individual player reports are done with the .anticheat player gm cmd.