Fix (module) False hits for Celestial Planetarium
This is a unqiue one. the light path leading to the circle centeral area seems to be false hitting for ignore z axis and tele plane, this is really the best solution I can come up with, since the map id is the whole instance and the zone and area id is the same, we lock it down to wmo area table entry by Tripple. Doing any form of ignore z-axis or telepane would not benefit the player due to with the mechanics dealing with the battle with the observer